expose isotyping api
