+++ docsrc/api.rst
@@ -23,3 +23,4 @@ Modules
 
     mutation_model
     utils
+    isotyping
